...G...... [ZBXNEXT-1078] added 'proc.mem' 5-th parameter (memtype) 'pmem' as a common parameter on Linux
git-svn-id: svn://svn.zabbix.com/branches/dev@49627 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBXNEXT-1078/src/libs/zbxsysinfo/freebsd/proc.c
@@ -336,7 +336,7 @@ int     PROC_MEM(AGENT_REQUEST *request, AGENT_RESULT *result)
 	else
 	{
 		if (ZBX_DO_AVG == do_task)
-			SET_DBL_RESULT(result, proccount == 0 ? 0.0 : (double)pct_size / (double)proccount);
+			SET_DBL_RESULT(result, proccount == 0 ? 0.0 : pct_size / (double)proccount);
 		else
 			SET_DBL_RESULT(result, pct_size);
 	}
